Pass and call the application from the controllers
Our idea is to convert an incoming OCCI request into an OpenStack
request, so that we can pass it downstream so as to be processed by
nova. However, we should call the final application in each of the
controlers instead of just modifying the request, in order to be able to
process the response from nova and "re-occi-fy" it.  In order to do so,
the controllers should receive the application.

 - This change modifies the BaseController so that in receives the app.
   If the controller returns a response (because it calls the
   application, gets the response and process it or by any other
   reason), that response will be used. If it returns None, the
   application will be called.
 - Also modifies the compute controller so that it implements the
   create_resources() function.