Pass down the openstack version to the controllers
OpenStack does some checks searching for a version string in the URL, so
we need to change the script path in the request to match what they're
looking for.

This change modifies the controllers so that they receive the version,
and adds a helpher method that prepares the request for passing it down
to the application.