Refactor the OCCI middleware tests
Create a new skeleton to ensure that all the controllers undergo the
same tests. Using a base test that does not set the "Accept" header on
the request (according to the OCCI specification this should render into
"text/plain"), and two subclasses that set the it to "text/plain" and
"text/occi". Afterwars, we can test each controller creating subclasses
of these three "base" tests.