Handle OpenStack API faults
We should start handling the OpenStack Faults, in case they exist. The
easiest way is checking if the status code belongs to one of the HTTP
error codes. If this is the case, we convert it back to a webob
exception and raise it so that it is eventually handled in our WSGI
middleware.