Preserve original request when calling OpenStack
This change creates a new request instead of modifying the original one
when it needs to be modified so as to access OpenStack. Before we were
modifying the original request (i.e. the path, body and content_type)
and this is wrong, since we cannot access those attributes after it has
been modified.