refactor compute controller index
This is part of the controller refactoring. The current desing is hard
to test, since each API action is done in a single method.  This
refactoring tries to split the construction of the OpenStack V2 api
requests from all the other logic, so that we can test it separatedly.
Moreover, this refactoring introduces new unit testing, instead of the
functional testing that is being done right now.

Change-Id: I94e5a6df995b5dd774422113d8ede7d1f808b500