refactor 'show' method from compute controller
This is part of the controller refactoring. The current desing is hard to test,
since each API action is done in a single method.  This refactoring tries to
split the construction of the OpenStack V2 api requests from all the other
logic, so that we can test it separatedly.  Moreover, this refactoring
introduces new unit testing, instead of the functional testing that is being
done right now.

Change-Id: I85be28a7f278e81b31c78e782c020056f5803d61