helpers: added a base class and compute helper
Part of the controller refactoring.

Introduce a new class 'OpenStackHelper' that will be used for making the
correct requests for the nova API and handlign the responses. This way
we can test individually each of the expected requests, and it will be
easier to switch from one version of the nova API to another by just
switching the helper class.

Change-Id: I5c80936a66c0121878dcefd42e3a92ed045e6a0c