Network refactoring
Refactor of the network handling. Now there are two networks: fixed and
floating. In order to allocate + assign an IP to a VM a link must be
created with network/floating as target. The pool from where the IP is
allocated can be specified with a mixin with term equal to the name of
the pool in openstack and schema
"http://schemas.openstack.org/network/floatingippool#".

Change-Id: I31d2769248679c844dfa901f675604958238af7f