Create OCCI and Controller network for neutron
Add controller for networks. It manages neutron or nova network depending of existing the parameter neutro_ooi_endpoint.
Create OCCI for network in ooi/openstack/network.py. It add a new mixin with one parameter (IP version), it can be used for the future to add more attributes.
Add tests for those changes.

Change-Id: I3c49ae9fcab9ac7f0adbaf559263a9d608793ce1