Update link handling in parser
Change the link structure returned by the parser to better follow the
OCCI 1.2 standard. The parser returns a dictionary where keys are the
link target kind type and values are lists with dictionaries for each
of the links of the links defined for that kind type. Those dictionaries
contain the target location (occi.core.target), the link id if defined,
and a dictionary with any extra attributes defined.

Change-Id: I24205044a6e2b670ceb98f3a0e9224637bf9fcf4