Increase attribute types
In order to do proper JSON rendering, ooi needs to to further refine the
Object attribute type into numbers, bools and strings. All attributes are
updated to include the type. Object attribute type is only kept for source
and targets in links.  Handling of attributes in network is simplified thanks
to this change.

Change-Id: If1c0e6b1e08139b369b3c68c4a12ff7244c3445f