Included OCCI IPReservation support
This is a complete change in which we have add the OCCI IPRESERVATION
class with the features provided by CESNET. We include: the controller
, the occi class, middleware changes, and tests to control it.
In addition, we have modified networklink to support this feature,
now networklink is identified by using only compute id and ipreservation
ip. Last, we add the documentation relate to its usage.

Change-Id: I52f80a9a7a39e00d1d7487001863586a551edf5f