Check if project id is None from headers
To check if the user was providing a scoped token we were relying on the
"HTTP_X_PROJECT_ID" environment variable, relying on a KeyError
exception when accessing the dictionary. However, the variable can be
empty, resulting in a dictionary key "HTTP_X_PROJECT_ID":None. As a
consequence, we were using None as the project ID, resulting in
misleading errors.

Closes-bug: #1678060
Change-Id: I1956fd99be5771371b1d3401e335528c5cc6455e