Publish and accept OCCI/1.1 as a valid version
Change d5ad90a fixed the publication of OCCI/1.1 as the supported OCCI
version, but it introduced a new bug that prevented users from passing
OCCI/1.1 in the user agent, forcing them to be OCCI/1.2. Even if the
OCCI spec says that: "The 'User-Agent' header MUST include the same
value (OCCI/X.Y) as advertised by the server" we should continue
accepting OCCI/1.1 as a valid version.

Change-Id: Iabcd8fda367876568b6f221d0ecc5ab2ffa06c90
Closes-Bug: #1683317