Use same environment for Neutron calls
We need to use the same environment (with the token, etc.) for making
the neutron calls, instead of just trying to obtain the token from the
headers (whose key lookup in the dictionary, by the way, was wrong). We
need to remove some keys from the environ, so that the new request is
created properly.

Change-Id: I86bedab940449a07780e288c03d4de789b8e695a